Make tests compile after abstracting over monad in async tests
