+++ octopus/src/main/scala/octopus/AsyncValidator.scala
@@ -2,15 +2,6 @@ package octopus
 
 import scala.language.higherKinds
 
-trait App[M[_]] {
-  def pure[A](a: A): M[A]
-  def map2[A, B, C](first: M[A], second: M[B])(combine: (A, B) => C): M[C]
-  def recover[A, B](app: M[A], f: Throwable => B): M[B]
-  def map[A, B](fa: M[A])(f: A => B): M[B] = {
-    map2(fa, pure(true))((a, _) => f(a))
-  }
-}
-
 trait AsyncValidator[M[_], T] {
   def validate(obj: T)(implicit app: App[M]): M[List[ValidationError]]
 }
